+++ src/main/java/br/com/reinodoce/mctiktok/client/ReinodoceClientBootstrap.java
@@ -3,6 +3,7 @@ package br.com.reinodoce.mctiktok.client;
 import br.com.reinodoce.mctiktok.command.ReinodoceCommandRegistrar;
 import br.com.reinodoce.mctiktok.platform.mc1201.Forge1201PlatformBridge;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.client.event.RegisterClientCommandsEvent;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 
@@ -19,7 +20,8 @@ public final class ReinodoceClientBootstrap {
             return;
         }
         SERVICE.initialize();
-        MinecraftForge.EVENT_BUS.addListener(ReinodoceCommandRegistrar::onRegisterClientCommands);
+        MinecraftForge.EVENT_BUS.addListener((RegisterClientCommandsEvent event) ->
+                ReinodoceCommandRegistrar.onRegisterClientCommands(event, SERVICE));
     }
 
     public static ReinodoceClientService service() {